refactor: move body components to each folder
